added user.service, also add check cookie validity so that we seperate concern and controller can focus on its task, and services under services folder can do the checking, fixed minor bugs

+++ server/services/session.service.js
@@ -1,6 +1,7 @@
 import prisma from "../prisma.js";
 
 async function createSession(userId) {
+  if (!userId) return null;
   const session = await prisma.session.create({
     data: {
       userPK: userId,
@@ -9,10 +10,24 @@ async function createSession(userId) {
   });
   return session;
 }
+
+async function createSessionTransaction(db, userId) {
+  if (!userId) return null;
+  const session = await db.session.create({
+    data: {
+      userPK: userId,
+      expiresAt: new Date(Date.now() + 1000 * 60 * 60 * 24),
+    },
+  });
+  return session;
+}
+
 async function deleteSession(sessionId) {
+  if (!sessionId) return null;
   await prisma.session.delete({ where: { sessionId: sessionId } });
 }
 async function updateSession(sessionId) {
+  if (!sessionId) return null;
   const session = await prisma.session.update({
     where: {
       sessionId: sessionId,
@@ -23,6 +38,7 @@ async function updateSession(sessionId) {
   });
 }
 async function getSession(sessionId) {
+  if (!sessionId) return null;
   const session = await prisma.session.findUnique({
     where: {
       sessionId: sessionId,
@@ -30,4 +46,10 @@ async function getSession(sessionId) {
   });
   return session;
 }
-export { createSession, deleteSession, updateSession, getSession };
+export {
+  createSession,
+  createSessionTransaction,
+  deleteSession,
+  updateSession,
+  getSession,
+};
